requirements.txt
streamlit>=1.39.0
langchain>=0.3.27
langchain-community>=0.3.0
langchain-openai>=0.2.0
tavily-python>=0.5.0
chromadb>=0.5.0
pydantic>=2.9.0
requests>=2.32.0
python-dotenv>=1.0.1
